Files rearranged, successful build.
